Make some type conversions explicit.
Fixes a bunch of MSVC warnings.

+++ src/Platform/Sdl2Application.cpp
@@ -254,13 +254,13 @@ Sdl2Application::Configuration::~Configuration() = default;
 Sdl2Application::InputEvent::Modifiers Sdl2Application::MouseEvent::modifiers() {
     if(modifiersLoaded) return _modifiers;
     modifiersLoaded = true;
-    return _modifiers = fixedModifiers(SDL_GetModState());
+    return _modifiers = fixedModifiers(Uint16(SDL_GetModState()));
 }
 
 Sdl2Application::InputEvent::Modifiers Sdl2Application::MouseMoveEvent::modifiers() {
     if(modifiersLoaded) return _modifiers;
     modifiersLoaded = true;
-    return _modifiers = fixedModifiers(SDL_GetModState());
+    return _modifiers = fixedModifiers(Uint16(SDL_GetModState()));
 }
 
 template class BasicScreen<Sdl2Application>;
